minor UI changes
