+++ index.html
@@ -10,10 +10,8 @@
     <meta name="keywords"
         content="cyber security, online security, data protection, network security, cyber attack, cyber defense, club, community, training, discussion">
     <title>Threat Hunters Club</title>
-    <link rel="shortcut icon"
-        href="https://ik.imagekit.io/fq9vykvp2/flaticon.png?ik-sdk-version=javascript-1.4.3&updatedAt=1675706897051"
-        type="image/x-icon">
-        
+    <link rel="shortcut icon" href="./src/assets/earth.png" type="image/x-icon">
+
     <link rel="stylesheet" href="index.css">
 </head>
