+++ src/components/Navbar/Header.css
@@ -18,6 +18,7 @@
     /* global 94%+ browsers support */
     background: linear-gradient(var(--gradient-text));
     -webkit-background-clip: text;
+    background-clip: text;
     -webkit-text-fill-color: transparent;
 }
 
@@ -25,6 +26,7 @@
 .thc__header {
     display: flex;
     overflow: hidden;
+    height: 60vh;
 }
 
 .thc__header-content {
@@ -56,16 +58,16 @@
 }
 
 .thc__header-img {
-    flex: 1; 
+    flex: 1;
     display: flex;
     justify-content: center;
     align-items: center;
 }
 
-.thc__header-img img {
+/* .thc__header-img img {
     width: 100%;
     height: 100%;
-}
+} */
 
 
 @media screen and (max-width:1050px) {
