+++ src/components/Navbar/Header.jsx
@@ -18,7 +18,7 @@ const Header = () => {
           minWidth: 100.0,
           scale: 1.0,
           scaleMobile: 0.6,
-          color: 0x0002A1,
+          color: 0x0002a1,
           size: 1.5,
           backgroundColor: 0x040c18,
         })
