+++ src/components/Navbar/Navbar.css
@@ -41,6 +41,7 @@
 .thc__navbar-links_container ul {
     display: flex;
     list-style-type: none;
+    align-items: center;
 }
 
 .thc__navbar-links_container ul li {
@@ -52,9 +53,9 @@
 .thc__navbar-links-btn {
     border: 1px solid var(--color-btns);
     outline: none;
-    border-radius: 10px;
-    padding: 5px 30px;
-    font-weight: 600;
+    border-radius: 5px;
+    padding: 5px 20px;
+    font-weight: 500;
     transition: background-color 0.1s ease-in-out;
 }
