+++ src/components/feature/whatcybersec.css
@@ -63,8 +63,8 @@
 
 .thc__whatCyberSec-features_container p {
     margin-top: 20px;
-    font-weight: 700;
-    font-size: 1.2rem;
+    font-weight: 500;
+    font-size: 1rem;
     text-align: left;
 }
 
@@ -92,7 +92,7 @@
 .thc__whatCyberSec-features_container {
     display: flex;
     flex-wrap: wrap;
-    margin-top: 2rem;
+    margin-top: 1rem;
 }
 
 /* Customizing feature component as per needs */
